+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.5.0] - 2022-06-16
+### Added
+- `Dates` and `Convert` moved from old `Date` classes to Java 8 time library 
+  including use of `LocalDate`, `LocalTime`, `LocalDateTime` and supporting classes.
+
 ## [0.4.5] - 2022-06-08
 ### Added
 - Conversion of string to UUID type.